Move Event Highlight Frame
Moves Event HighlightFrame and Pane from Console to general Cbus Swing
